Moved out jsonschema + sampling to separate files. Changed exposed "sample" to "sampling" and in general structured sampling.js in a comparable way to pioneer-utils sampling.ts
